% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_confidence_draws.R
\name{add_mean_confidence_draws}
\alias{add_mean_confidence_draws}
\title{Obtain posterior draws of mean confidence separately for each possible stimulus}
\usage{
add_mean_confidence_draws(newdata, object, ...)
}
\arguments{
\item{newdata}{Data frame from which to generate posterior predictions}

\item{object}{The \code{brms} model with the \code{metad} family}

\item{...}{Additional parameters passed to \code{tidybayes::epred_draws}}
}
\value{
a tibble containing posterior draws of mean confidence with the following
columns:
.row: the row of \code{newdata}
.chain, .iteration, .draw: identifiers for the posterior sample
stimulus: indicator for stimulus presence
.epred: the predicted mean confidence
}
\description{
Obtain posterior draws of mean confidence separately for each possible stimulus
}
