% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacognitive_bias_draws.R
\name{add_metacognitive_bias_draws}
\alias{add_metacognitive_bias_draws}
\title{Get draws of \code{meta-delta}, an index of metacognitive bias.}
\usage{
add_metacognitive_bias_draws(newdata, object, ..., by_response = TRUE)
}
\arguments{
\item{newdata}{Data frame from which to generate estimates}

\item{object}{the \code{brms} model with the \code{metad} family}

\item{...}{additional parameters to pass to \code{tidybayes::linpred_draws}}

\item{by_response}{If \code{TRUE}, generate separate estimates for each response.}
}
\description{
Get draws of \code{meta-delta}, an index of metacognitive bias.
}
