% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_draws.R
\name{add_roc1_draws}
\alias{add_roc1_draws}
\title{Obtain posterior draws of the pseudo type 1 receiver operating characteristic (ROC) curve.}
\usage{
add_roc1_draws(newdata, object, ..., bounds = FALSE)
}
\arguments{
\item{newdata}{Data frame from which to generate posterior predictions}

\item{object}{the \code{brms} model with the \code{metad} family}

\item{...}{Additional parameters passed to \code{tidybayes::epred_draws}}

\item{bounds}{If \code{TRUE}, include the endpoints of the ROC at \verb{(0, 0)} and \verb{(1, 1)}.}
}
\value{
a tibble containing posterior draws of the pseudo type 1 ROC with the following
columns:
.row: the row of \code{newdata}
.chain, .iteration, .draw: identifiers for the posterior sample
joint_response: the combined type 1 / type 2 response (in 1:2*K for K confidence levels)
response: the type 1 response for perceived stimulus presence
confidence: the type 2 confidence response
p_fa: the cumulative probability of a 'present'/'old' response for stimulus==0
p_hit: the cumulative probability of a 'present'/'old' response for stimulus==1
}
\description{
Obtain posterior draws of the pseudo type 1 receiver operating characteristic (ROC) curve.
}
