% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRatio.R
\name{joint_response}
\alias{joint_response}
\alias{response_probabilities}
\alias{metad_aggregate}
\title{Convert a type 1 response (0 or 1) and a corresponding confidence level
(between 1 and K) into a joint response between 1 and 2*K, with outer
values reflecting confident responses and intermediate values reflecting
uncertainty.}
\usage{
joint_response(response, confidence, K)

response_probabilities(x)

metad_aggregate(data, ..., .response = "N", K = NULL)
}
\arguments{
\item{response}{type 1 response (0 or 1)}

\item{confidence}{type 2 response/confidence rating (1:K)}

\item{K}{The number of confidence levels in \code{data}. If \code{NULL}, this is estimated from \code{data}.}

\item{x}{A vector of counts of joint type 1/type 2 responses}

\item{data}{the tibble to aggregate}

\item{...}{grouping columns in the tibble}

\item{.response}{the name of the column containing trial counts}
}
\value{
A vector of response probabilities

A tibble with one row per combination of the variables in \code{...},
and another column named by the value of \code{.response} containing trial counts.
For K confidence levels, this will be an \verb{N x K*4} matrix, such that the
columns represent:
\verb{[N(stimulus==0, confidence==K), ..., N(stimulus==0, confidence==1), N(stimulus==0, confidence==1), ..., N(stimulus==0, confidence==K), N(stimulus==1, confidence==K), ..., N(stimulus==1, confidence==1), N(stimulus==1, confidence==1), ..., N(stimulus==1, confidence==K)]}
}
\description{
Convert a type 1 response (0 or 1) and a corresponding confidence level
(between 1 and K) into a joint response between 1 and 2*K, with outer
values reflecting confident responses and intermediate values reflecting
uncertainty.

For a vector of counts of joint type 1/type 2 responses,
compute the corresponding probabilities.

Aggregate \code{data} by columns \code{response}, \code{confidence},
and any other variables in \code{...}.
}
