% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{cov_matrix}
\alias{cov_matrix}
\title{Generate a covariance matrix.}
\usage{
cov_matrix(S, OMEGA)
}
\arguments{
\item{S}{a vector of standard deviations}

\item{OMEGA}{a correlation matrix}
}
\value{
a \verb{[N x N]} covariance matrix, where \code{N = length(S)}.
}
\description{
Generate a covariance matrix.
}
