% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRatio.R
\name{fit_metad}
\alias{fit_metad}
\title{Fit the metad' model using the \code{brms} package}
\usage{
fit_metad(
  formula,
  data,
  ...,
  aggregate = TRUE,
  K = NULL,
  distribution = "normal",
  metac_absolute = TRUE,
  stanvars = NULL
)
}
\arguments{
\item{formula}{A model formula for some or all parameters of the \code{metad} brms family.
To display all parameter names for a model with \code{K} confidence levels, use \code{metad(K)}.}

\item{data}{A tibble containing the data to fit the model. If \code{aggregate}==TRUE,
\code{data} should have one row per observation. If \code{aggregate}==FALSE, it should be aggregated
to have one row per cell of the design matrix, with joint
type 1/type 2 response counts in a matrix column.}

\item{...}{Additional parameters passed to the \code{brm} function.}

\item{aggregate}{If \code{TRUE}, automatically aggregate \code{data} by the variables included in \code{formula}.
Otherwise, \code{data} should already be aggregated.}

\item{K}{The number of confidence levels. By default, this is estimated from the data.}

\item{distribution}{The noise distribution to use for the signal detection model}

\item{metac_absolute}{If \code{TRUE}, fix the type 2 criterion to be equal to the type 1 criterion.
Otherwise, equate the criteria relatively such that metac/metadprime = c/dprime.}

\item{stanvars}{Additional \code{stanvars} to pass to the model code, for example to define an alternative
distribution or a custom model prior.}
}
\description{
Fit the metad' model using the \code{brms} package
}
