% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metad_family.R
\name{get_dist}
\alias{get_dist}
\title{Get the R function for the model's underlying distribution functions}
\usage{
get_dist(model, fun = "lcdf")
}
\arguments{
\item{model}{The \code{brms} model to get distribution functions for}

\item{fun}{The distribution function to return.}
}
\value{
An R function with the name \verb{distribution_\{fun\}}.
Will throw an error if this function does not exist
}
\description{
Get the R function for the model's underlying distribution functions
}
