% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metad_family.R
\name{get_metac}
\alias{get_metac}
\title{Get the parameterization of \code{meta_c} in \code{model}}
\usage{
get_metac(model)
}
\arguments{
\item{model}{the \code{brms} model}
}
\value{
A character vector, either \code{"absolute"} or \code{"relative"}.
}
\description{
Get the parameterization of \code{meta_c} in \code{model}
}
