% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRatio.R
\name{joint_response}
\alias{joint_response}
\title{Convert a type 1 response (0 or 1) and a corresponding confidence level
(between 1 and K) into a joint response between 1 and 2*K, with outer
values reflecting confident responses and intermediate values reflecting
uncertainty.}
\usage{
joint_response(response, confidence, K)
}
\arguments{
\item{response}{type 1 response (0 or 1)}

\item{confidence}{type 2 response/confidence rating (1:K)}

\item{K}{maximum confidence level}
}
\description{
Convert a type 1 response (0 or 1) and a corresponding confidence level
(between 1 and K) into a joint response between 1 and 2*K, with outer
values reflecting confident responses and intermediate values reflecting
uncertainty.
}
