% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metad_family.R
\name{log_lik_metad}
\alias{log_lik_metad}
\title{Generate a function to calculate the log likelihood of the metad' model}
\usage{
log_lik_metad(i, prep)
}
\arguments{
\item{i}{an observation index}

\item{prep}{an object containing the data and model draws}
}
\value{
A \verb{[D x K*4]} array containing posterior samples of
the joint probability of a type 1/type 2 response,
where \code{D} is the number of posterior draws,
\code{N} is the number of rows in the data, and
\code{K} is the number of confidence levels.
}
\description{
Generate a function to calculate the log likelihood of the metad' model
}
