% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metad_family.R
\name{lp_metad}
\alias{lp_metad}
\title{Calculate the log probability simplex of the metad' model}
\usage{
lp_metad(i, prep)
}
\arguments{
\item{i}{an observation index}

\item{prep}{an object containing the data and model draws}
}
\value{
A vector of joint type 1/type 2 response probabilties
for observation \code{i} in \code{prep}
}
\description{
Calculate the log probability simplex of the metad' model
}
