% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{cov_matrix}
\alias{cov_matrix}
\alias{corr_matrix}
\alias{rmatrixnorm}
\title{Generate a covariance matrix.}
\usage{
cov_matrix(S, OMEGA)

corr_matrix(r, nrow = 2)

rmatrixnorm(mu, L_sigma_1, L_sigma_2)
}
\arguments{
\item{S}{a vector of standard deviations}

\item{OMEGA}{a correlation matrix}

\item{r}{the correlation to fill in the matrix off-diagonals}

\item{nrow}{the number of rows (and columns) of the resulting matrix}

\item{mu}{a matrix of means}

\item{L_sigma_1}{the Cholesky-decomposed covariance matrix for the rows}

\item{L_sigma_2}{the Cholesky-decomposed covariance matrix for the columns}
}
\value{
a \verb{[N x N]} covariance matrix, where \code{N = length(S)}.

An \verb{[nrow x nrow]} matrix with values along the diagonal equal
to \code{1} and values off of the diagonal equal to \code{r}

A single sample from a matrix-normal distribution with mean
\code{mu} (a matrix), row-wise covariances \code{sigma_1}, and column-wise
covariances \code{sigma_2}, where \code{L_sigma_1} and \code{L_sigma_2} are the
Cholesky-decomposed covariance matrices
}
\description{
Generate a covariance matrix.

Generate an \verb{[nrow x nrow]} correlation matrix
with all off-diagonal values equal to r

Sample from a matrix-normal distribution
}
