% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_confidence_draws.R
\name{mean_confidence_draws}
\alias{mean_confidence_draws}
\title{Obtain posterior draws of mean confidence separately for each possible stimulus}
\usage{
mean_confidence_draws(
  object,
  newdata,
  ...,
  by_stimulus = TRUE,
  by_response = TRUE
)
}
\arguments{
\item{object}{the \code{brms} model with the \code{metad} family}

\item{newdata}{Data frame from which to generate posterior predictions}

\item{...}{Additional arguments to tidybayes::epred_draws}

\item{by_stimulus}{If TRUE, predict mean confidence separately by stimulus}

\item{by_response}{If TRUE, predict mean confidence separately by response
Otherwise, predict mean confidence averaging across stimuli.}
}
\value{
a tibble containing posterior draws of mean confidence with the following
columns:
.row: the row of \code{newdata}
.chain, .iteration, .draw: identifiers for the posterior sample
stimulus: indicator for stimulus presence
.epred: the predicted mean confidence
}
\description{
Obtain posterior draws of mean confidence separately for each possible stimulus
}
