% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metad_family.R
\name{metad}
\alias{metad}
\title{Generate a brms family for the metad' model with K confidence levels}
\usage{
metad(K, distribution = "normal", metac_absolute = TRUE)
}
\arguments{
\item{K}{The number of confidence levels}

\item{distribution}{The noise distribution to use for the signal detection model}

\item{metac_absolute}{If \code{TRUE}, fix the type 2 criterion to be equal to the type 1 criterion.
Otherwise, equate the criteria relatively such that \code{metac/metadprime = c/dprime}.}
}
\value{
A \code{brms} family for the metad' model with K confidence levels
}
\description{
Generate a brms family for the metad' model with K confidence levels
}
