% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRatio.R
\name{metad_aggregate}
\alias{metad_aggregate}
\title{Aggregate \code{data} by columns \code{response}, \code{confidence},
and any other variables in \code{...}.}
\usage{
metad_aggregate(data, ..., .response = "N", K = NULL)
}
\arguments{
\item{data}{the tibble to aggregate}

\item{...}{grouping columns in the tibble}

\item{.response}{the name of the column containing trial counts}

\item{K}{The number of confidence levels in \code{data}. If \code{NULL}, this is estimated from \code{data}.}
}
\value{
A tibble with one row per combination of the variables in \code{...},
and another column named by the value of \code{.response} containing trial counts.
For K confidence levels, this will be an \verb{N x K*4} matrix, such that the
columns represent:
\verb{[N(stimulus==0, confidence==K), ..., N(stimulus==0, confidence==1), N(stimulus==0, confidence==1), ..., N(stimulus==0, confidence==K), N(stimulus==1, confidence==K), ..., N(stimulus==1, confidence==1), N(stimulus==1, confidence==1), ..., N(stimulus==1, confidence==K)]}
}
\description{
Aggregate \code{data} by columns \code{response}, \code{confidence},
and any other variables in \code{...}.
}
