% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metad_draws.R
\name{metad_draws}
\alias{metad_draws}
\alias{add_metad_draws}
\alias{metad_rvars}
\alias{add_metad_rvars}
\title{Obtain posterior draws of meta-d' model parameters}
\usage{
metad_draws(object, newdata, ..., pivot_longer = FALSE)

add_metad_draws(newdata, object, ..., pivot_longer = FALSE)

metad_rvars(object, newdata, ..., pivot_longer = FALSE)

add_metad_rvars(newdata, object, pivot_longer = FALSE)
}
\arguments{
\item{object}{The \code{brms} model with the \code{metad} family}

\item{newdata}{A data frame from which to generate posterior predictions}

\item{...}{Additional arguments passed to \link[tidybayes:add_predicted_draws]{tidybayes::add_linpred_draws} or
\link[tidybayes:add_predicted_rvars]{tidybayes::add_linpred_rvars}}

\item{pivot_longer}{Return the draws in long format?
\itemize{
\item if \code{TRUE}, resulting data frame has one row per posterior draw per model parameter
\item if \code{FALSE} (default), resulting data frame has one row per posterior draw
}}
}
\value{
a tibble containing posterior draws of model parameters with the
following columns:
\itemize{
\item \code{.row}: the row of \code{newdata}
\item \code{.chain}, \code{.iteration}, \code{.draw}: for \code{metad_draws}, identifiers for the posterior sample
\item \code{.variable}, \code{.value}: if \code{pivot_longer=TRUE}, \code{.variable} identifies different meta-d' model parameters and \code{.value} stores posterior samples
\item \code{M}, \code{dprime}, \code{c}, \code{meta_dprime}, \code{meta_c}, \verb{meta_c2_0_<k>}, \verb{meta_c2_1_<k>}: if \code{pivot_longer=FALSE}, posterior samples of all meta-d' model parameters
}
}
\description{
Given a data frame and a meta-d' model, adds estimates of all
model parameters. For \code{metad_draws} and \code{add_metad_draws}, parameters are
returned in a tidy tibble with one row per posterior draw. For
\code{metad_rvars} and \code{add_metad_rvars}, parameters are returned as
\link[posterior:rvar]{posterior::rvar}s, with one row per row in \code{newdata}.
}
\examples{
# running few iterations so example runs quickly, use more in practice
m <- fit_metad(N ~ 1, sim_metad(), chains = 1, iter = 500)
newdata <- tidyr::tibble(.row = 1)

# obtain model parameters (wide format)
metad_draws(m, newdata)
add_metad_draws(newdata, m)

# obtain model parameters (long format)
metad_draws(m, newdata, pivot_longer = TRUE)
add_metad_draws(newdata, m, pivot_longer = TRUE)

# obtain model parameters (wide format, posterior::rvar)
metad_rvars(m, newdata)
add_metad_rvars(newdata, m)

# obtain model parameters (long format, posterior::rvar)
metad_rvars(m, newdata, pivot_longer = TRUE)
add_metad_rvars(newdata, m, pivot_longer = TRUE)

}
