% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metad_family.R
\name{metad_pmf}
\alias{metad_pmf}
\title{Generate (log) probability simplex over the joint type 1/type 2 responses}
\usage{
metad_pmf(
  stimulus,
  d_prime,
  c,
  meta_d_prime,
  meta_c,
  meta_c2_0,
  meta_c2_1,
  lcdf = normal_lcdf,
  lccdf = normal_lccdf,
  log = FALSE
)
}
\arguments{
\item{stimulus}{the stimulus (0 or 1)}

\item{d_prime}{the type 1 sensitivity}

\item{c}{the type 1 response criterion}

\item{meta_d_prime}{the type 2 sensitivity}

\item{meta_c}{the type 1 criteriom for generating confidence ratings}

\item{meta_c2_0}{the type 2 response criteria for \code{"0"} responses}

\item{meta_c2_1}{the type 2 response criteria for \code{"1"} responses}

\item{lcdf}{The log cumulative distribution function for the underlying distribution in the metad' model.
By default, uses the normal distribution with a standard deviation of \code{1}.}

\item{lccdf}{The log complement cumulative distribution function for the underlying distribution in the metad' model.
By default, uses the normal distribution with a standard deviation of \code{1}.}

\item{log}{if TRUE, return log probabilities instead of probabilities}
}
\description{
Generate (log) probability simplex over the joint type 1/type 2 responses
}
