% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metad_family.R
\name{metad_stancode}
\alias{metad_stancode}
\title{Generate Stan code for the meta-d' model}
\usage{
metad_stancode(K, distribution = "normal", metac_absolute = TRUE)
}
\arguments{
\item{K}{The number of confidence levels}

\item{distribution}{The noise distribution to use. Should be a parameter-free
distribution, i.e., one that is mean-centered without additional variance/shape parameters.
If the distribution is not already available in stan, you must additionally provide two
functions to Stan (one for \verb{<distribution>_lcdf} and one for \verb{<distribution>_lccdf}).}

\item{metac_absolute}{Should the type 2 criterion (metac) be fixed to the absolute type 1 criterion (c)?
If \code{TRUE}, the model will set \code{metac = c}. Otherwise, it will set \code{metac = M * c}, such that
the type 2 criterion is \emph{relatively} equal to the type 1 criterion
(i.e., \code{meta_c/meta_d_prime = c/d_prime})}
}
\value{
A single string containing Stan code defining the likelihood for the metad' model
with \code{K} confidence levels, signal distributed according to the distribution \code{distribution},
and where metac = c if metac_absolute is true, and metac = M*c otherwise.
}
\description{
Generate Stan code for the meta-d' model
}
