% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRatio-package.R
\docType{package}
\name{mRatio-package}
\alias{mRatio}
\alias{mRatio-package}
\title{mRatio: Fit the meta-d' model of confidence ratings using `brms`}
\description{
Implementation of a `brms` custom family for the meta-d\' model (Maniscalco & Lau, 2012).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://metacoglab.github.io/mRatio/}
  \item \url{https://github.com/metacoglab/mRatio}
  \item Report bugs at \url{https://github.com/metacoglab/mRatio/issues}
}

}
\author{
\strong{Maintainer}: Kevin O'Neill \email{kevin.o'neill@ucl.ac.uk}

}
\keyword{internal}
