% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metad_family.R
\name{posterior_epred_metad}
\alias{posterior_epred_metad}
\title{Generate posterior predictions for the metad' model}
\usage{
posterior_epred_metad(prep)
}
\arguments{
\item{prep}{an object containing the data and model draws}
}
\value{
A \verb{[D x N x K*4]} array containing posterior samples of
the joint probability of a type 1/type 2 response,
where \code{D} is the number of posterior draws,
\code{N} is the number of rows in the data, and
\code{K} is the number of confidence levels.
}
\description{
Generate posterior predictions for the metad' model
}
