% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metad_family.R
\name{posterior_predict_metad}
\alias{posterior_predict_metad}
\title{Simulate posterior predictions from the metad' model}
\usage{
posterior_predict_metad(i, prep, ...)
}
\arguments{
\item{i}{an observation index}

\item{prep}{an object containing the data and model draws}

\item{...}{Additional arguments. Not currently used.}
}
\value{
A \verb{[D x K*4]} array containing posterior samples of
counts of joint type 1/type 2 responses,
where \code{D} is the number of posterior draws,
\code{N} is the number of rows in the data, and
\code{K} is the number of confidence levels.
}
\description{
Simulate posterior predictions from the metad' model
}
