% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRatio.R
\name{response_probabilities}
\alias{response_probabilities}
\title{For a vector of counts of joint type 1/type 2 responses,
compute the corresponding probabilities.}
\usage{
response_probabilities(x)
}
\arguments{
\item{x}{A vector of counts of joint type 1/type 2 responses}
}
\value{
A vector of response probabilities
}
\description{
For a vector of counts of joint type 1/type 2 responses,
compute the corresponding probabilities.
}
