% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{rmatrixnorm}
\alias{rmatrixnorm}
\title{Sample from a matrix-normal distribution}
\usage{
rmatrixnorm(mu, L_sigma_1, L_sigma_2)
}
\arguments{
\item{mu}{a matrix of means}

\item{L_sigma_1}{the Cholesky-decomposed covariance matrix for the rows}

\item{L_sigma_2}{the Cholesky-decomposed covariance matrix for the columns}
}
\value{
A single sample from a matrix-normal distribution with mean
\code{mu} (a matrix), row-wise covariances \code{sigma_1}, and column-wise
covariances \code{sigma_2}, where \code{L_sigma_1} and \code{L_sigma_2} are the
Cholesky-decomposed covariance matrices
}
\description{
Sample from a matrix-normal distribution
}
