% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRatio.R
\name{to_signed}
\alias{to_signed}
\title{Convert binary variable \code{x} from \verb{\{0, 1\}} to \verb{\{-1, 1\}}}
\usage{
to_signed(x)
}
\arguments{
\item{x}{a binary variable interpretable as \code{0} or \code{1}}
}
\value{
\code{1} if \code{x==1}, else \code{-1}
}
\description{
Convert binary variable \code{x} from \verb{\{0, 1\}} to \verb{\{-1, 1\}}
}
