% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacognitive_bias_draws.R
\name{metacognitive_bias}
\alias{metacognitive_bias}
\title{Given the distances between successive confidence thresholds,
calculate the average of the cumulative distances to 0.}
\usage{
metacognitive_bias(...)
}
\arguments{
\item{...}{a series of distances between confidence thresholds}
}
\description{
Given the distances between successive confidence thresholds,
calculate the average of the cumulative distances to 0.
}
