% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRatio-package.R
\docType{package}
\name{mRatio-package}
\alias{mRatio}
\alias{mRatio-package}
\title{mRatio: Fit the meta-d' model of confidence ratings using `brms`}
\description{
Implementation of a `brms` custom family for the meta-d\' model (Maniscalco & Lau, 2012).
}
\author{
\strong{Maintainer}: Kevin O'Neill \email{kevin.o'neill@ucl.ac.uk}

}
\keyword{internal}
