% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_draws.R
\name{roc2_draws}
\alias{roc2_draws}
\title{Obtain posterior draws of the response-specific type 2 receiver operating characteristic (ROC) curves.}
\usage{
roc2_draws(object, newdata, ..., bounds = FALSE)
}
\arguments{
\item{object}{the \code{brms} model with the \code{metad} family}

\item{newdata}{Data frame from which to generate posterior predictions}

\item{...}{Additional parameters passed to \code{tidybayes::epred_draws}}

\item{bounds}{If \code{TRUE}, include the endpoints of the ROC at \verb{(0, 0)} and \verb{(1, 1)}.}
}
\value{
a tibble containing posterior draws of the response-specific type 2 ROCs with the following
columns:
.row: the row of \code{newdata}
.chain, .iteration, .draw: identifiers for the posterior sample
response: the type 1 response for perceived stimulus presence
confidence: the type 2 confidence response
p_fa2: the cumulative probability of an incorrect but confident response
p_hit2: the cumulative probability of a correct and confident response
}
\description{
Obtain posterior draws of the response-specific type 2 receiver operating characteristic (ROC) curves.
}
